+++ data-transfer/src/main/java/it/eng/datatransfer/model/TransferProcess.java
@@ -61,6 +61,8 @@ public class TransferProcess extends AbstractTransferMessage {
 	private String agreementId;
 	@JsonIgnore
 	private String callbackAddress;
+	@JsonIgnore
+	private String fileId;
 	
 	// determins which role the connector is for that contract negotiation (consumer or provider)
     @JsonIgnore
@@ -108,13 +110,13 @@ public class TransferProcess extends AbstractTransferMessage {
 			return this;
 		}
 
-		@JsonProperty((DSpaceConstants.DSPACE_PROVIDER_PID))
+		@JsonProperty(DSpaceConstants.DSPACE_PROVIDER_PID)
 		public Builder providerPid(String providerPid) {
 			message.providerPid = providerPid;
 			return this;
 		}
 		
-		@JsonProperty((DSpaceConstants.DSPACE_STATE))
+		@JsonProperty(DSpaceConstants.DSPACE_STATE)
 		public Builder state(TransferState state) {
 			message.state = state;
 			return this;
@@ -134,7 +136,7 @@ public class TransferProcess extends AbstractTransferMessage {
         	message.format = format;
         	return this;
         }
-
+		
 		@JsonProperty("createdBy")
 		public Builder createdBy(String createdBy) {
 			message.createdBy = createdBy;
@@ -162,6 +164,12 @@ public class TransferProcess extends AbstractTransferMessage {
 		@JsonProperty(DSpaceConstants.CALLBACK_ADDRESS)
 		public Builder callbackAddress(String callbackAddress) {
 			message.callbackAddress = callbackAddress;
+			return this;
+		}
+		
+		@JsonProperty("fileId")
+		public Builder fileId(String fileId) {
+			message.fileId = fileId;
 			return this;
 		}
 
@@ -207,6 +215,7 @@ public class TransferProcess extends AbstractTransferMessage {
 				.format(this.format)
 				.state(newTransferState)
 				.role(this.role)
+				.fileId(this.fileId)
 				// no need to modify audit fields???
 				.createdBy(this.createdBy)
 				.lastModifiedBy(this.lastModifiedBy)