+++ data-transfer/src/main/java/it/eng/datatransfer/rest/api/RestArtifactController.java
@@ -3,15 +3,16 @@ package it.eng.datatransfer.rest.api;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
 import it.eng.datatransfer.service.api.RestArtifactService;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 
 @RestController
@@ -31,24 +32,30 @@ public class RestArtifactController {
 	 * 
 	 * @param authorization
 	 * @param transactionId Base64.urlEncoded(consumerPid|providerPid) from TransferProcess message
-	 * @param artifactId artifactId
 	 * @param jsonBody
 	 * @return
 	 */
-    @PostMapping(path = "/{transactionId}/{artifactId}")
-    protected ResponseEntity<String> getArtifact(@RequestHeader(required = false) String authorization,
+    @RequestMapping(path = "/{transactionId}", method = { RequestMethod.GET, RequestMethod.POST })
+    protected ResponseEntity<String> getArtifact(HttpServletResponse response,
+    												@RequestHeader(required = false) String authorization,
 										    		@PathVariable String transactionId,                                       
-										    		@PathVariable String artifactId, 
-                                                  @RequestBody(required = false) JsonNode jsonBody) {
-    	log.info("Accessing artifact with id {}", artifactId);
+										    		@RequestBody(required = false) JsonNode jsonBody) {
     
-		String response = restArtifactService.getArtifact(transactionId, artifactId, jsonBody);
-		
-        return ResponseEntity.ok()
+		String rr = restArtifactService.getArtifact(transactionId, jsonBody);
+		return ResponseEntity.ok()
                 .contentType(MediaType.APPLICATION_JSON)
-                .body(response);
-
+                .body(rr);
+		/*
+		 * Uncomment this part when TransferProcess.State.INITIAL is done
+		 * this logic requires to have fileId present in TransferPRocess
+		try (ServletOutputStream outputStream = response.getOutputStream()) {
+			GridFsResource attachment = restArtifactService.streamAttachment(transactionId);
+			response.setHeader("Content-Disposition", "attachment;filename=\"" + attachment.getFilename() + "\"");
+			response.addHeader("Content-type", attachment.getContentType());
+			IOUtils.copy(attachment.getInputStream(), response.getOutputStream());
+		} catch (IOException e) {
+			response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
+		}
+		*/
     }
-    
-  
 }