+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 23-10-2024
+
+### Added
+
+ - Upload and list artifacts that will be shared as dataset
+ - When verifying agreement additional check (policyEnforcement for agreement exists) added
+ - New CatalogErrorAPIException that translates to HTTP 400 response
+ 
 ## [0.1.1] - 20-10-2024
 
 ### Added